requirements.txt
altgraph==0.17.3
anyio==3.6.1
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
future==0.18.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
numpy==1.23.3
openai==0.23.1
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.0
pandas-stubs==1.5.0.220926
pefile==2022.5.30
pexels-api-py==0.0.5
pycparser==2.21
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
PySocks==1.7.1
python-amazon-paapi==5.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-wordpress-xmlrpc==2.3
pytz==2022.4
pywin32-ctypes==0.2.0
requests==2.28.1
rfc3986==1.5.0
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
termcolor==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
types-pytz==2022.2.1.0
typing_extensions==4.3.0
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
youtube-search-python==1.6.6
